utilities/workflow-maker/pyproject.toml
[project]
name = "workflow-maker"
version = "0.1.0"
description = "Convert markdown agent files to workflow JSON"
requires-python = ">=3.12"
dependencies = [
    "pyyaml>=6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
